+++ tools/src/main/java/org/esa/cci/sst/tool/Tool.java
@@ -203,13 +203,13 @@ public abstract class Tool {
         }
 
         options.addOption(createOption("h", "help", null,
-                                       "displays this help."));
+                                       "Displays this help."));
         options.addOption(createOption("v", "version", null,
-                                       "display the version of this program and exit."));
+                                       "Displays the version of this program and exits."));
         options.addOption(createOption("c", "config", "FILE",
-                                       "read configuration from given FILE."));
+                                       "Reads a configuration (key-value pairs) from given FILE."));
         options.addOption(createOption("e", "errors", null,
-                                       "dumps full error stack trace."));
+                                       "Dumps a full error stack trace."));
         options.addOption(createOption("l", "logLevel", "LEVEL",
                                        String.format("sets the logging level. Must be one of %s. Use level '%s' to also output diagnostics. The default value is '%s'.",
                                                      Arrays.toString(LogLevel.values()), LogLevel.all, DEFAULT_LOG_LEVEL)));